big refactor encapsulate the host inventory in its own module
